refactor: move utility functions from utils to helper for better organization
